% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{bootstrap_vb}
\alias{bootstrap_vb}
\title{Fit Parameteric Bootstrap for Variational Bayes}
\usage{
bootstrap_vb(method, data, B = 10)
}
\arguments{
\item{method}{[character] The path to stan file containing model fitting
code.}

\item{data}{[list] A list of data to input in the data{} field of the stan
file.}

\item{B}{[integer] The number of bootstrap replicates to compute.}
}
\value{
result [list] A list containing beta and theta fields, which is an
  array similar to what is output by stan(), except columns are now bootstrap
  replicates instead of sampling iterations
}
\description{
Fit Parameteric Bootstrap for Variational Bayes
}

