% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{clear_tmp}
\alias{clear_tmp}
\title{Clear all files in tempdir()}
\usage{
clear_tmp()
}
\description{
Hack used to avoid crashing https://groups.google.com/forum/#!category-topic/stan-users/general/Bg8HhyB6qG4
}

