% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{error_histograms}
\alias{error_histograms}
\title{Errors histogram}
\usage{
error_histograms(plot_data, facet_terms = NULL, n_bins = 75, alpha = 0.7,
  colors = c("#d95f02", "#7570b3"))
}
\arguments{
\item{plot_data}{[data.frame] The data used to plot the error between truth
vs. estimate across all dimensions. See the output of
melt_reshaped_samples().}

\item{facet_terms}{[character vector] The columns on which to facet_grid the
plot.}

\item{n_bins}{[int] The number of bins in each histogram panel. Defaults to 75.}

\item{alpha}{[numeric] The alpha transparency for the different factors.}

\item{colors}{[character vector] The colors to use for each factor.}
}
\value{
hist_plot [ggplot] The ggplot object showing error histograms across
  factors and simulation configurations.
}
\description{
Plot the histograms of errors associated with the scatterplots from the NMF fits.
}

