% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_model}
\alias{fit_model}
\title{Fit a generic NMF Stan Model}
\usage{
fit_model(y, model_opts = list())
}
\arguments{
\item{y}{[matrix] The data on which to fit the NMF model.}

\item{model_opts}{[list] A partially filled list of model fitting options.
Unspecified options will be passed into merge_model_opts().}
}
\value{
result [stan object] The fitted stan object.
}
\description{
This wraps vb() and stan() in the STAN package to let us run either approach
using a single command.
}

