% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{match_matrix}
\alias{match_matrix}
\title{Identify a permutation that aligns rows of two matrices}
\usage{
match_matrix(X, Z)
}
\arguments{
\item{X}{[numeric matrix] A matrix whose rows we want to align with
X.}

\item{Z}{[numeric matrix] A matrix whose rows we want to align with
Z.}
}
\value{
pi_result [vector] A permutation such that X[pi_result, ] = Z
(ideally).
}
\description{
One way to get confidence intervals for mixtures on the simplex is
to first specify the cluster labels for each component, and then
study each of these histograms on their own. This is different
from, say, smoothing the mixture distribution and identifying
modes.
}
\details{
The approach taken here is to find the two rows with maximal
correlation and put that in the required permutation. Then, remove
those rows and repeat.
}
\examples{
X <- matrix(rnorm(100, mean = 4) ^ 2, 20, 5)
pi <- sample(1:20)
Z <- X[pi, ] + matrix(rnorm(100), 20, 5)
pi_hat <- match_matrix(X, Z)
cbind(pi_hat, pi)
}

