% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{melt_reshaped_samples}
\alias{melt_reshaped_samples}
\title{Melt reshaped samples}
\usage{
melt_reshaped_samples(samples)
}
\arguments{
\item{samples}{[data.frame] The wide samples data, usually the output of
reshape_all_samples.}
}
\value{
melted_samples [data.frame] The same data as samples, but with
  different factor dimensions all stacked.
}
\description{
While for the scatter / contours figures, it's useful to have the dimensions
as separate columns, we'll also want to the melted data when computing
explicit errors. This takes the output of reshaped_... and melts it so that
it's appropriate for histogramming the errors.
}

