% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{merge_model_opts}
\alias{merge_model_opts}
\title{Merge Default Modeling Parameters}
\usage{
merge_model_opts(opts = list())
}
\arguments{
\item{opts}{[list] A partially filled list of options, to fill in with
defaults.}
}
\value{
opts [list] The version of opts with defaults filled in.
}
\description{
This lets us run models with partially specified options.
}

