% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{merge_nmf_opts}
\alias{merge_nmf_opts}
\title{Merge Default NMF options}
\usage{
merge_nmf_opts(opts = list())
}
\arguments{
\item{opts}{[list] A partially filled list of options, to fill in with
defaults.}
}
\value{
opts [list] The version of opts with defaults filled in.
}
\description{
Merge Default NMF options
}

