% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{nmf_posterior_means}
\alias{nmf_posterior_means}
\title{Wrapper to get posterior means in NMF}
\usage{
nmf_posterior_means(samples)
}
\arguments{
\item{samples}{The output of a call to vb() using the nmf code}
}
\value{
result [list] A list with the posterior means as matrices, with
  factors (k) as columns
}
\description{
Wrapper to get posterior means in NMF
}

