% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{nmf_sim}
\alias{nmf_sim}
\title{Simulate NMF Data (optionally zero inflated)}
\usage{
nmf_sim(opts)
}
\arguments{
\item{opts}{[list] A list containing parameters for simulation. Any options
that are not specified will be filled in with defaults, according to
merge_nmf_defaults().}
}
\value{
A list with the latent thetas, betas, mask, and observed Y.
}
\description{
To facilitate simulation across many parameters, it's useful to have a single
function to generate all the quantities of interest.
}

