% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{reshape_all_samples}
\alias{reshape_all_samples}
\title{Wrapper for reshape_samples, applying too many fits}
\usage{
reshape_all_samples(fits, config_path, param, dims, sim_seed = 1112017)
}
\arguments{
\item{fits}{[list of paths] A list of paths to the fitted STAN samples. They
will be loaded and the associated samples will be extracted.}

\item{config_path}{[character] The path to the configuration JSON containing
all the experiment information for these fits. It is assumed that the names
in fits include the (necessarily numerical) IDs specified in this JSON
object.}

\item{param}{[character] The name of the parameter to extract data for.}

\item{dims}{[character vector] The names to use for the rows and columns of
the parameter matrix.}

\item{sim_seed}{[integer] The seed used in simulating the parameter of
interest. This is necessary because we don't read the true values anywhere,
we regenerate the parameter values.}
}
\value{
reshaped [data.table] The same kind of object as the output of
  reshape_samples(), except over experiments with multiple configurations.
}
\description{
We might want to extract the true and fitted values across several nmf
experiments. Each experiment individually would require calling
reshape_samples.
}

