% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{reshape_samples}
\alias{reshape_samples}
\title{Reshape fitted scores for easy plotting}
\usage{
reshape_samples(samples, truth, dims)
}
\arguments{
\item{samples}{[array] The samples generated by STAN; e.g. obtained by
fit${param}. Two dimensions are for the actual matrix, the third are from
sampling iterations.}

\item{truth}{[matrix] The truth parameter values, known because we are
working from a simulation.}

\item{dims}{[character vector] The names to use for the rows and columns of
the parameter matrix.}
}
\value{
reshaped [data.table] The melted scores, with truth along with samples.
}
\description{
Both thetas and betas in NMF are reshaped using very similar code, so we
might as well combine it explicitly in one function.
}

