% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{score_means}
\alias{score_means}
\title{Get Score Means}
\usage{
score_means(scores, grouping_vars)
}
\arguments{
\item{scores}{[data.frame] The data.frame giving the scores for true and
simulated data.}

\item{grouping_vars}{[character vector] The names of columns on which to
group. Everything else will be averaged over.}
}
\value{
scores [data.frame] The original data.frame with the value_1 and
  value_2 columns averaged depending on grouping variables.
}
\description{
This groups the main scores matrix by sample index, and average posterior
samples.
}

