% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scores_contours}
\alias{scores_contours}
\title{Plot Contours and Associated Coordinate}
\usage{
scores_contours(plot_data, plot_opts)
}
\arguments{
\item{plot_data}{[data.frame] The data frame that contains the sampled
coordinates for the contours, along with the true positions (which will be
indicated with text).}

\item{plot_opts}{[list] A list specifying plotting appearance. Pretty much
the same as the usual ggcontours plot_opts, except an extra option for plot
limits.}
}
\value{
[list] A list containing two ggplot2 objects, for the faceted and
  unfaceted scores plots.
}
\description{
This is a wrapper of ggcontours in the ggscaffold package that shows the
coordinate associated with point clouds.
}

