% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_from_params}
\alias{sim_from_params}
\title{Simulate (ZI)-NMF data from theta, beta, and inflation parameters}
\usage{
sim_from_params(theta, beta, zero_inf_prob = 0)
}
\arguments{
\item{theta}{[matrix] An N x K matrix of scores}

\item{beta}{[matrix] A V x K matrix of factors}

\item{zero_inf_prob}{[numeric] The amount of zero inflation; i.e., the
proportion in [0, 1] that will be set to zero.}
}
\value{
y [matrix] The N x V matrix of observed (potentially zero-inflated
  counts).
}
\description{
Simulate (ZI)-NMF data from theta, beta, and inflation parameters
}

