% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{write_configs}
\alias{write_configs}
\title{Write experiment configurations files}
\usage{
write_configs(sim_factors, model_factors, n_batches = 50,
  config_path = "config.json", output_dir = "./")
}
\arguments{
\item{factors}{[list] A list of one-dimensional factors on which we compute
the full cross-product.}

\item{path}{[string] The path to which to save the configurations JSON.}
}
\description{
This generates the JSON file on which all the experiments will be based. It
just creates a JSON will the full factorial combinations of input options.
}
\examples{
sim_factors <- list(
  "N" = c(50, 100, 200),
  "P" = c(75, 125),
  "zero_inf_prob" = c(0, 0.2, 0.5, 0.8)
)
model_factors <- list(
  "inference" = c("gibbs", "vb"),
  "method" = c("zinf_nmf", "nmf")
)
#write_configs(sim_factors, model_factors)
}

